Refactor notifications to launch a notification handler
With this change we no can launch external urls. For all I know this
doesn't work without this trick (an intent in the middle).

I also applied a transparent theme to the activity to avoid seeing the
launched activity briefly before opening external urls.

This is based on my experience working on "Share To Clipboard".