Fix adding members on some devices.
Networking is not allowed on the main thread, and on some devices with strict
mode on, even the creation of the http handler is enough to trigger an
exception (i.e even if not used from the thread).

This moves even the creation to a thread which fixes the issue.