Revert: Import: always import, don't try to find duplicates based on UID
This is somehow a revert of 536bef9815555487fe03edf42302db6c39b5522f.

It was initially implemented as a workaround for #24, but having
improved all the clients to deal with weird UIDs and matured a lot since
then, I believe this is no longer an issue.

Would have to keep a close eye for regressions.