Revert "Add a reentrant lock to our entity data store."
I'm pretty sure this was causing the recent issues with tasks and events
disappearing as described in #97.

Regardless of this, it's not actually needed as SQLite on Android is
already thread-safe. The SQLite locks some users were experiencing were
probably fixed in 26ea8900a200d3b4c3accc730c03928b966af63d

This reverts commit 9ed172e23c57aaa3ee7d2d2f43c1e7c85a2c8c7b.