EtebaseCache: fix crash when removing and re-adding (sometimes).
Essentially the problem was that when we were removing the account, the cache
directory was removed. Though then when readding the account, we would get the
cache handler from our handler cache, which would already be init, so it'd
assume the directory already exists. This was only relevant if there were
network errors because otherwise the sync would also happen it the background
which would create the missing dirs anyway.

The solution: remove the object from the handler cache when removing account.